backend/requirements.txt
blinker==1.6.2
click==8.1.4
dnspython==2.3.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.7.0
pymongo==4.4.0
python-dotenv==1.0.0
python-ldap==3.4.3
Werkzeug==2.3.6